% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Dataset Description}
\format{
A data frame with specific format
}
\usage{
data(dataset)
}
\description{
This is a dataset used for demonstration purposes in the sanketphonetictranslator package.
}
\keyword{datasets}
